backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.10.1
pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
